refactor colorKit frm class to functions
